refactor: minor cleanups
