+++ lib/widgets/common/carousel/carousel_style_registry.dart
@@ -1,6 +1,5 @@
 import 'package:anymex/models/Media/media.dart';
-import 'package:anymex/widgets/common/big_carousel_v2.dart';
-import 'package:anymex/widgets/common/carousel/big_carousel_classic.dart';
+import 'package:anymex/widgets/common/carousel/themes/big_carousel_classic.dart';
 import 'package:anymex/widgets/common/carousel/carousel_types.dart';
 import 'package:flutter/widgets.dart';
 
@@ -32,12 +31,6 @@ class CarouselStyleRegistry {
       description: 'Full-width hero banner with compact details and quick synopsis.',
       builder: _buildClassic,
     ),
-    CarouselStyleDefinition(
-      id: 'cinematic',
-      name: 'Cinematic',
-      description: 'Centered showcase cards with expanded visuals and smoother focus.',
-      builder: _buildCinematic,
-    ),
   ];
 
   static int normalizeIndex(int index) {
@@ -66,13 +59,5 @@ class CarouselStyleRegistry {
     required CarouselType carouselType,
   }) {
     return BigCarouselClassic(key: key, data: data, carouselType: carouselType);
-  }
-
-  static Widget _buildCinematic({
-    Key? key,
-    required List<Media> data,
-    required CarouselType carouselType,
-  }) {
-    return BigCarouselV2(key: key, data: data, carouselType: carouselType);
   }
 }
