+++ lib/widgets/common/carousel/themes/big_carousel_classic.dart
